# docs/openapi/cliser-data-node.openapi.current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLISER DATA NODE API",
    "version": "1.0.0",
    "description": "HTTP API do CLISER DATA NODE.\n\nAPI Version: v1\nBase Path: /api/v1"
  },
  "paths": {
...
    "/api/v1/lifecycle/gc": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Garbage Collect",
        "operationId": "garbage_collect_api_v1_lifecycle_gc_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/lifecycle/rebuild-refcounts": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Rebuild Refcounts",
        "operationId": "rebuild_refcounts_api_v1_lifecycle_rebuild_refcounts_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/lifecycle/integrity-check": {
      "post": {
        "tags": [
          "lifecycle"
        ],
        "summary": "Integrity Check",
        "operationId": "integrity_check_api_v1_lifecycle_integrity_check_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace}/objects": {
      "get": {
        "tags": [
          "objects"
        ],
        "summary": "List Objects",
        "operationId": "list_objects_api_v1_namespaces__namespace__objects_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectListResponse"
                },
                "examples": {
                  "ObjectListResponse": {
                    "$ref": "#/components/examples/ObjectListResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "objects"
        ],
        "summary": "Create Object",
        "operationId": "create_object_api_v1_namespaces__namespace__objects_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            },
            "name": "Idempotency-Key",
            "in": "header"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectCreateRequest"
              },
              "examples": {
                "ObjectCreate": {
                  "$ref": "#/components/examples/ObjectCreateRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectCreateResponse"
                },
                "examples": {
                  "ObjectCreateResponse": {
                    "$ref": "#/components/examples/ObjectCreateResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/namespaces/{namespace}/objects/{object_id}/download": {
      "get": {
        "tags": [
          "download"
        ],
        "summary": "Download de objeto",
        "operationId": "download_object_api_v1_namespaces__namespace__objects__object_id__download_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Conteúdo binário do objeto.",
            "headers": {
              "Content-Length": {
                "description": "Tamanho do conteúdo em bytes.",
                "schema": {
                  "type": "integer"
                }
              },
              "X-Object-ID": {
                "description": "Identificador do objeto.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Content-Hash": {
                "description": "Hash SHA-256 do conteúdo.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Recupera o conteúdo binário de um objeto armazenado no namespace informado."
      }
    },
...
    "/api/v1/namespaces/{namespace}/objects/{object_id}/manifest": {
      "get": {
        "tags": [
          "manifests"
        ],
        "summary": "Get Manifest",
        "operationId": "get_manifest_api_v1_namespaces__namespace__objects__object_id__manifest_get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            },
            "name": "object_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestResponse"
                },
                "examples": {
                  "ManifestResponse": {
                    "$ref": "#/components/examples/ManifestResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ManifestBlockResponse": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "index",
          "block_id",
          "size"
        ],
        "title": "ManifestBlockResponse"
      },
      "ManifestResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "total_size": {
            "type": "integer",
            "title": "Total Size"
          },
          "block_size": {
            "type": "integer",
            "title": "Block Size"
          },
          "block_count": {
            "type": "integer",
            "title": "Block Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/ManifestBlockResponse"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "total_size",
          "block_size",
          "block_count",
          "created_at",
          "status",
          "blocks"
        ],
        "title": "ManifestResponse"
      },
...
      "ObjectCreateRequest": {
        "properties": {
          "data": {
            "type": "string",
            "minLength": 1,
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ObjectCreateRequest"
      },
      "ObjectCreateResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "content_hash",
          "status"
        ],
        "title": "ObjectCreateResponse"
      },
...
      "ObjectListItem": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "size",
          "status"
        ],
        "title": "ObjectListItem"
      },
      "ObjectListResponse": {
        "properties": {
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/ObjectListItem"
            },
            "type": "array",
            "title": "Objects"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "namespace",
          "objects",
          "count"
        ],
        "title": "ObjectListResponse"
      },
...
      "APIError": {
        "title": "APIError",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "OBJECT_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Objeto não encontrado"
              },
              "request_id": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "req_01JEXAMPLE"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "credential",
        "description": "Credencial de API enviada no cabeçalho Authorization como Bearer <credential>."
      }
    },
    "examples": {
...
      "ObjectCreateRequest": {
        "value": {
          "data": "Q0xJU0VSIERBVEEgTk9ERQ=="
        },
        "summary": "Objeto em Base64",
        "description": "O campo data deve conter Base64 válido."
      },
      "ObjectCreateResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "size": 16,
          "content_hash": "sha256:01JEXAMPLE",
          "status": "ACTIVE"
        },
        "summary": "Objeto criado"
      },
      "ObjectListResponse": {
        "value": {
          "namespace": "documents",
          "objects": [
            {
              "object_id": "obj_01JEXAMPLE",
              "namespace": "documents",
              "size": 16,
              "created_at": "2026-09-14T12:00:00Z",
              "status": "ACTIVE"
            }
          ],
          "count": 1
        },
        "summary": "Lista de objetos"
      },
...
      "ManifestResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "total_size": 524288,
          "block_size": 262144,
          "block_count": 2,
          "created_at": "2026-09-15T15:00:00Z",
          "status": "ACTIVE",
          "blocks": [
            {
              "index": 0,
              "block_id": "blk_01JEXAMPLE0",
              "size": 262144
            },
            {
              "index": 1,
              "block_id": "blk_01JEXAMPLE1",
              "size": 262144
            }
          ]
        },
        "summary": "Manifesto do objeto"
      },
...
      "ErrorValidation": {
        "value": {
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Request validation failed.",
            "request_id": "req_01JEXAMPLE",
            "details": {
              "errors": [
                {
                  "type": "value_error",
                  "loc": [
                    "body",
                    "namespace"
                  ],
                  "msg": "field required"
                }
              ]
            }
          }
        },
        "summary": "Erro de validação"
      },
...
    }
  },
...
}